Update snapshot and Rename MockAppComponent
