+++ src/__mocks__/MockAppComponent.js
